added build persistence
