+++ .circleci/config.yml
@@ -28,21 +28,19 @@ jobs:
             python3 main.py patch
 
       - run:
-          name: "Run test suite"
-          no_output_timeout: 10m
+          name: "Generate Testing Binary"
+          no_output_timeout: 20m
           command: |
-            cd terraform-provider-aws
-            go test -c ./internal/service/s3
+            python3 main.py build -s s3
 
   run-tests:
     machine:
       image: ubuntu-2004:202101-01
     resource_class: medium
     steps:
-      - checkout
       - go/install:
             version: '1.19'
-      - run: git submodule update --init --recursive
+
       - run:
           name: "Installing prerequisites"
           command: |
@@ -51,28 +49,11 @@ jobs:
             pip3 install -r requirements.txt
 
       - run:
-          name: "Configuring repos for tests"
-          command: |
-            cd terraform-provider-aws && go mod vendor
-            cd ../
-            python3 main.py patch
-
-      - run:
           name: "Run test suite"
           no_output_timeout: 30m
           command: |
-            docker-compose up -d
-            export TF_ACC=1            
-            export AWS_ALTERNATE_ACCESS_KEY_ID=test
-            export AWS_ALTERNATE_SECRET_ACCESS_KEY=test
-            export AWS_ALTERNATE_REGION=us-east-2
-            export AWS_DEFAULT_REGION=eu-west-1
-            export AWS_ACCESS_KEY_ID=test
-            export AWS_SECRET_ACCESS_KEY=test
-            cd terraform-provider-aws
-            go test ./internal/service/s3 -test.count 1 -test.v -test.timeout 60m -parallel 4
-#            export PYTHONUNBUFFERED=1
-#            pytest terraform-provider-aws/internal/service/s3 -s -v --ls-start
+            export PYTHONUNBUFFERED=1
+            pytest terraform-provider-aws/internal/service/s3 -s -v --ls-start
 
   generate:
     machine:
@@ -103,8 +84,8 @@ workflows:
         - equal: [ build-new, << pipeline.git.branch >> ]
         - equal: [ pytest-plugin, << pipeline.git.branch >> ]
     jobs:
-      - generate
-      - test:
+      - generate-test-binary
+      - run-tests:
           requires:
-            - generate
+            - generate-test-binary
