--- constants.py
@@ -1,7 +1,0 @@
-BASE_PATH = 'terraform-provider-aws'
-SERVICE_BASE_PATH = './internal/service'
-
-BIN_PATH = 'test-bin'
-
-PATCH_PATH = 'etc'
-PATCH_FILES = ['001-hardcode-endpoint.patch']
